+++ src/DatePicker/__tests__/DatePicker-test.js
@@ -15,7 +15,7 @@ describe('DatePicker', () => {
     const handleSelect = jest.fn()
     const instance = TestUtils.renderIntoDocument(<DatePicker date="2016-01-01" onSelect={handleSelect} />)
     const container = findDOMNode(instance)
-    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+    TestUtils.Simulate.click(container)
     TestUtils.Simulate.click(document.querySelector('tbody button'))
     expect(handleSelect).toBeCalled()
   })