+++ src/DatePicker/__tests__/DateRange-test.js
@@ -9,7 +9,7 @@ describe('DateRange', () => {
     const handleSelect = jest.fn()
     const instance = TestUtils.renderIntoDocument(<DateRange start="2016-01-01" onSelect={handleSelect} />)
     const container = findDOMNode(instance)
-    TestUtils.Simulate.click(container.querySelector('.bfd-dropdown__toggle'))
+    TestUtils.Simulate.click(container.querySelector('.bfd-datepicker'))
     TestUtils.Simulate.click(document.querySelector('tbody button'))
     expect(handleSelect).toBeCalled()
   })